+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/aio/_text_analytics_client_async.py
@@ -861,7 +861,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
     @validate_multiapi_args(
         version_method_added="v3.1",
         args_mapping={
-            "2022-10-01-preview": ["fhir_version", "document_type", "autodetect_default_language"],
+            "2022-10-01-preview": ["fhir_version", "document_type"],
             "2022-05-01": ["display_name"]
         }
     )
@@ -869,7 +869,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         self,
         documents: Union[List[str], List[TextDocumentInput], List[Dict[str, str]]],
         *,
-        autodetect_default_language: Optional[str] = None,
         continuation_token: Optional[str] = None,
         disable_service_logs: Optional[bool] = None,
         display_name: Optional[str] = None,
@@ -910,8 +909,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             2022-10-01-preview and newer). If not set, uses "en" for English as default.
             Per-document language will take precedence over whole batch language.
             See https://aka.ms/talangs for supported languages in Language API.
-        :keyword str autodetect_default_language: Default/fallback language to use for documents requesting
-            automatic language detection.
         :keyword str display_name: An optional display name to set for the requested analysis.
         :keyword str string_index_type: Specifies the method used to interpret string offsets.
             Can be one of 'UnicodeCodePoint' (default), 'Utf16CodeUnit', or 'TextElement_v8'.
@@ -951,7 +948,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         .. versionadded:: 2022-05-01
             The *display_name* keyword argument.
         .. versionadded:: 2022-10-01-preview
-            The *fhir_version*, *document_type*, and *autodetect_default_language* keyword arguments.
+            The *fhir_version* and *document_type* keyword arguments.
 
         .. admonition:: Example:
 
@@ -1008,7 +1005,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                         body=models.AnalyzeTextJobsInput(
                             analysis_input=input_docs,
                             display_name=display_name,
-                            default_language=autodetect_default_language,
                             tasks=[
                                 models.HealthcareLROTask(
                                     task_name="0",
@@ -1093,9 +1089,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
     @validate_multiapi_args(
         version_method_added="v3.1",
         custom_wrapper=check_for_unsupported_actions_types,
-        args_mapping={
-            "2022-10-01-preview": ["autodetect_default_language"],
-        }
     )
     async def begin_analyze_actions(
         self,
@@ -1116,7 +1109,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             ]
         ],
         *,
-        autodetect_default_language: Optional[str] = None,
         continuation_token: Optional[str] = None,
         display_name: Optional[str] = None,
         language: Optional[str] = None,
@@ -1177,8 +1169,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             2022-10-01-preview and newer). If not set, uses "en" for English as default.
             Per-document language will take precedence over whole batch language.
             See https://aka.ms/talangs for supported languages in Language API.
-        :keyword str autodetect_default_language: Default/fallback language to use for documents requesting
-            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword int polling_interval: Waiting time between two polls for LRO operations
             if no Retry-After header is present. Defaults to 5 seconds.
@@ -1215,7 +1205,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         .. versionadded:: 2022-10-01-preview
             The *ExtractSummaryAction* and *AbstractSummaryAction* input options and the corresponding
             *ExtractSummaryResult* and *AbstractSummaryResult* result objects.
-            The *autodetect_default_language* keyword argument.
 
         .. admonition:: Example:
 
@@ -1287,7 +1276,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                         body=models.AnalyzeTextJobsInput(
                             analysis_input=docs,
                             display_name=display_name,
-                            default_language=autodetect_default_language,
                             tasks=generated_tasks
                         ),
                         cls=response_cls,
@@ -1363,9 +1351,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
     @distributed_trace_async
     @validate_multiapi_args(
         version_method_added="2022-05-01",
-        args_mapping={
-            "2022-10-01-preview": ["autodetect_default_language"],
-        }
     )
     async def begin_recognize_custom_entities(
         self,
@@ -1373,7 +1358,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         project_name: str,
         deployment_name: str,
         *,
-        autodetect_default_language: Optional[str] = None,
         continuation_token: Optional[str] = None,
         disable_service_logs: Optional[bool] = None,
         display_name: Optional[str] = None,
@@ -1403,8 +1387,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             2022-10-01-preview and newer). If not set, uses "en" for English as default.
             Per-document language will take precedence over whole batch language.
             See https://aka.ms/talangs for supported languages in Language API.
-        :keyword str autodetect_default_language: Default/fallback language to use for documents requesting
-            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword bool disable_service_logs: If set to true, you opt-out of having your text input
             logged on the service side for troubleshooting. By default, the Language service logs your
@@ -1436,8 +1418,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
 
         .. versionadded:: 2022-05-01
             The *begin_recognize_custom_entities* client method.
-        .. versionadded:: 2022-10-01-preview
-            The *autodetect_default_language* keyword argument.
 
         .. admonition:: Example:
 
@@ -1484,7 +1464,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                             disable_service_logs=disable_service_logs
                         )
                     ],
-                    autodetect_default_language=autodetect_default_language,
                     display_name=display_name,
                     show_stats=show_stats,
                     language=language,
@@ -1500,9 +1479,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
     @distributed_trace_async
     @validate_multiapi_args(
         version_method_added="2022-05-01",
-        args_mapping={
-            "2022-10-01-preview": ["autodetect_default_language"],
-        }
     )
     async def begin_single_label_classify(
         self,
@@ -1510,7 +1486,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         project_name: str,
         deployment_name: str,
         *,
-        autodetect_default_language: Optional[str] = None,
         continuation_token: Optional[str] = None,
         disable_service_logs: Optional[bool] = None,
         display_name: Optional[str] = None,
@@ -1539,8 +1514,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             2022-10-01-preview and newer). If not set, uses "en" for English as default.
             Per-document language will take precedence over whole batch language.
             See https://aka.ms/talangs for supported languages in Language API.
-        :keyword str autodetect_default_language: Default/fallback language to use for documents requesting
-            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword bool disable_service_logs: If set to true, you opt-out of having your text input
             logged on the service side for troubleshooting. By default, the Language service logs your
@@ -1568,8 +1541,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
 
         .. versionadded:: 2022-05-01
             The *begin_single_label_classify* client method.
-        .. versionadded:: 2022-10-01-preview
-            The *autodetect_default_language* keyword argument.
 
         .. admonition:: Example:
 
@@ -1615,7 +1586,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                         )
                     ],
                     polling_interval=polling_interval_arg,
-                    autodetect_default_language=autodetect_default_language,
                     display_name=display_name,
                     show_stats=show_stats,
                     language=language,
@@ -1630,9 +1600,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
     @distributed_trace_async
     @validate_multiapi_args(
         version_method_added="2022-05-01",
-        args_mapping={
-            "2022-10-01-preview": ["autodetect_default_language"],
-        }
     )
     async def begin_multi_label_classify(
         self,
@@ -1640,7 +1607,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         project_name: str,
         deployment_name: str,
         *,
-        autodetect_default_language: Optional[str] = None,
         continuation_token: Optional[str] = None,
         disable_service_logs: Optional[bool] = None,
         display_name: Optional[str] = None,
@@ -1669,8 +1635,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             2022-10-01-preview and newer). If not set, uses "en" for English as default.
             Per-document language will take precedence over whole batch language.
             See https://aka.ms/talangs for supported languages in Language API.
-        :keyword str autodetect_default_language: Default/fallback language to use for documents requesting
-            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword bool disable_service_logs: If set to true, you opt-out of having your text input
             logged on the service side for troubleshooting. By default, the Language service logs your
@@ -1698,8 +1662,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
 
         .. versionadded:: 2022-05-01
             The *begin_multi_label_classify* client method.
-        .. versionadded:: 2022-10-01-preview
-            The *autodetect_default_language* keyword argument.
 
         .. admonition:: Example:
 
@@ -1745,7 +1707,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                         )
                     ],
                     polling_interval=polling_interval_arg,
-                    autodetect_default_language=autodetect_default_language,
                     display_name=display_name,
                     show_stats=show_stats,
                     language=language,
@@ -1867,7 +1828,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         self,
         documents: Union[List[str], List[TextDocumentInput], List[Dict[str, str]]],
         *,
-        autodetect_default_language: Optional[str] = None,
         continuation_token: Optional[str] = None,
         disable_service_logs: Optional[bool] = None,
         display_name: Optional[str] = None,
@@ -1898,8 +1858,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             2022-10-01-preview and newer). If not set, uses "en" for English as default.
             Per-document language will take precedence over whole batch language.
             See https://aka.ms/talangs for supported languages in Language API.
-        :keyword str autodetect_default_language: Default/fallback language to use for documents requesting
-            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword Optional[int] max_sentence_count: Maximum number of sentences to return. Defaults to 3.
         :keyword Optional[str] order_by:  Possible values include: "Offset", "Rank". Default value: "Offset".
@@ -1979,7 +1937,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                         )
                     ],
                     polling_interval=polling_interval_arg,
-                    autodetect_default_language=autodetect_default_language,
                     display_name=display_name,
                     show_stats=show_stats,
                     language=language,
@@ -1999,7 +1956,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         self,
         documents: Union[List[str], List[TextDocumentInput], List[Dict[str, str]]],
         *,
-        autodetect_default_language: Optional[str] = None,
         continuation_token: Optional[str] = None,
         disable_service_logs: Optional[bool] = None,
         display_name: Optional[str] = None,
@@ -2032,8 +1988,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             2022-10-01-preview and newer). If not set, uses "en" for English as default.
             Per-document language will take precedence over whole batch language.
             See https://aka.ms/talangs for supported languages in Language API.
-        :keyword str autodetect_default_language: Default/fallback language to use for documents requesting
-            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword Optional[int] sentence_count: It controls the approximate number of sentences in the output summaries.
         :keyword Optional[str] model_version: The model version to use for the analysis.
@@ -2111,7 +2065,6 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                         )
                     ],
                     polling_interval=polling_interval_arg,
-                    autodetect_default_language=autodetect_default_language,
                     display_name=display_name,
                     show_stats=show_stats,
                     language=language,